+++ NEWS.md
@@ -1,3 +1,29 @@
+## Version 0.2.0
+* Removed the subdirectories `wave` and `core` from the REAPER build, as they're
+only used for file IO.
+* Made REAPER patches to ensure that all diagnostic messages are CRAN-compliant,
+and added a `GetDiagnostics()` procedure to capture various internally used 
+measures with `reaper_wrap()` instead of writing them to disk.
+* Added several `output` options to `reaper()` and `reaper_bulk()`: `gci_cand`
+for information about glottal closure instants candidates, `probs` for
+various pseudo-probabilities relating to voicing, and `resids` for residual 
+waveforms for debugging.
+* Removed the `delete` parameters from `read_epochs_out()` and 
+`read_pitch_out()`. 
+* Refactored `reaper_bulk()` to allow for new outputs smoothly. Also added
+parameter `force_list_output` to `reaper()` as part of this.
+* Added `hirst2pass_f0min` and `hirst2pass_f0max` arguments to `reaper_bulk()`
+for user controlling the pitch floor and ceiling values for first pass.
+* Added `fileExtension` parameter to `reaper2emuDB()` for selecting
+the extension of SSFF files.
+* Various other minor fixes.
+
+## Version 0.1.1
+* Fixed time domain bug when estimating epochs with `start` times other than `0`.
+* Made minimal REAPER patches to fix CRAN check warnings, as well as a fix to
+the `StdoutSilencer` in the C++ wrapper.
+* Suppress diagnostic messages from REAPER by default.
+
 ## Version 0.1.0
 * Previous versions of the library attempted to install the `reaper` command
 line tool with CMake, and functions would call REAPER from the command line